--- a/benchmarks/Vectorization/polybench/MLIRPolybench2mm.mlir
+++ b/benchmarks/Vectorization/polybench/MLIRPolybench2mm.mlir
@@ -20,7 +20,7 @@
 //
 //===----------------------------------------------------------------------===//
 
-func.func @kernel_2mm_init_array(%arg0: i32, %arg1: i32, %arg2: i32, %arg3: i32, %arg4: memref<?xf64>, %arg5: memref<?xf64>, %arg6: memref<?x?xf64>, %arg7: memref<?x?xf64>, %arg8: memref<?x?xf64>, %arg9: memref<?x?xf64>) {
+func.func @polybench_2mm_init_array(%arg0: i32, %arg1: i32, %arg2: i32, %arg3: i32, %arg4: memref<?xf64>, %arg5: memref<?xf64>, %arg6: memref<?x?xf64>, %arg7: memref<?x?xf64>, %arg8: memref<?x?xf64>, %arg9: memref<?x?xf64>) {
   %c0 = arith.constant 0 : index
   %c1 = arith.constant 1 : index
   %c2_i32 = arith.constant 2 : i32
@@ -94,7 +94,7 @@ func.func @kernel_2mm_init_array(%arg0: i32, %arg1: i32, %arg2: i32, %arg3: i32,
   return
 }
 
-func.func @kernel_2mm(%arg0: i32, %arg1: i32, %arg2: i32, %arg3: i32, %arg4: f64, %arg5: f64, %arg6: memref<?x?xf64>, %arg7: memref<?x?xf64>, %arg8: memref<?x?xf64>, %arg9: memref<?x?xf64>, %arg10: memref<?x?xf64>) {
+func.func @polybench_2mm_kernel(%arg0: i32, %arg1: i32, %arg2: i32, %arg3: i32, %arg4: f64, %arg5: f64, %arg6: memref<?x?xf64>, %arg7: memref<?x?xf64>, %arg8: memref<?x?xf64>, %arg9: memref<?x?xf64>, %arg10: memref<?x?xf64>) {
   %cst = arith.constant 0.000000e+00 : f64
   %0 = arith.index_cast %arg2 : i32 to index
   %1 = arith.index_cast %arg3 : i32 to index