--- a/benchmarks/Vectorization/polybench/MLIRPolybench3mm.mlir
+++ b/benchmarks/Vectorization/polybench/MLIRPolybench3mm.mlir
@@ -1,3 +1,25 @@
+//===- MLIRPolybench3mm.mlir ----------------------------------------------===//
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+//===----------------------------------------------------------------------===//
+//
+// This file provides initialization and kernel functions for the 3mm Polybench
+// benchmark. The MLIR code is generated with Polygeist and modified manually to
+// run on different dataset sizes.
+//
+//===----------------------------------------------------------------------===//
+
 func.func @kernel_3mm_init_array(%arg0: i32, %arg1: i32, %arg2: i32, %arg3: i32, %arg4: i32, %arg5: memref<?x?xf64>, %arg6: memref<?x?xf64>, %arg7: memref<?x?xf64>, %arg8: memref<?x?xf64>) {
   %c0 = arith.constant 0 : index
   %c1 = arith.constant 1 : index