--- a/benchmarks/Vectorization/polybench/MLIRPolybenchAdi.mlir
+++ b/benchmarks/Vectorization/polybench/MLIRPolybenchAdi.mlir
@@ -43,7 +43,7 @@ func.func  @adi_init_array(%arg0: i32, %arg1: memref<?x?xf64>) {
   return
 }
 
-func.func  @adi(%arg0: i32, %arg1: i32, %arg2: memref<?x?xf64>, %arg3: memref<?x?xf64>, %arg4: memref<?x?xf64>, %arg5: memref<?x?xf64>) {
+func.func  @adi_kernel(%arg0: i32, %arg1: i32, %arg2: memref<?x?xf64>, %arg3: memref<?x?xf64>, %arg4: memref<?x?xf64>, %arg5: memref<?x?xf64>) {
   %cst = arith.constant 1.000000e+00 : f64
   %cst_0 = arith.constant 2.000000e+00 : f64
   %cst_1 = arith.constant 0.000000e+00 : f64