… benchmarks/Vectorization/CMakeLists.txt
… benchmarks/Vectorization/polybench/CMakeLists.txt
… benchmarks/Vectorization/polybench/MLIRPolybench3mmBenchmark.cpp
… benchmarks/Vectorization/polybench/MLIRPolybenchAdiBenchmark.cpp
… benchmarks/Vectorization/polybench/MLIRPolybenchAtax.mlir
… benchmarks/Vectorization/polybench/MLIRPolybenchAtaxBenchmark.cpp
… benchmarks/Vectorization/polybench/MLIRPolybenchBicg.mlir
… benchmarks/Vectorization/polybench/MLIRPolybenchBicgBenchmark.cpp
… benchmarks/Vectorization/polybench/MLIRPolybenchCholesky.mlir
… benchmarks/Vectorization/polybench/MLIRPolybenchCholeskyBenchmark.cpp
… benchmarks/Vectorization/polybench/MLIRPolybenchCorrelation.mlir
… benchmarks/Vectorization/polybench/MLIRPolybenchCorrelationBenchmark.cpp
… benchmarks/Vectorization/polybench/MLIRPolybenchCovariance.mlir
… benchmarks/Vectorization/polybench/MLIRPolybenchCovarianceBenchmark.cpp
… benchmarks/Vectorization/polybench/MLIRPolybenchDeriche.mlir
… benchmarks/Vectorization/polybench/MLIRPolybenchDericheBenchmark.cpp
… benchmarks/Vectorization/polybench/MLIRPolybenchDoitgen.mlir
… benchmarks/Vectorization/polybench/MLIRPolybenchDoitgenBenchmark.cpp
… benchmarks/Vectorization/polybench/MLIRPolybenchDurbin.mlir
… benchmarks/Vectorization/polybench/MLIRPolybenchDurbinBenchmark.cpp
… benchmarks/Vectorization/polybench/MLIRPolybenchFdtd2D.mlir
… benchmarks/Vectorization/polybench/MLIRPolybenchFdtd2DBenchmark.cpp
… benchmarks/Vectorization/polybench/MLIRPolybenchFloydWarshall.mlir
… benchmarks/Vectorization/polybench/MLIRPolybenchFloydWarshallBenchmark.cpp
… benchmarks/Vectorization/polybench/MLIRPolybenchGemm.mlir
… benchmarks/Vectorization/polybench/MLIRPolybenchGemmBenchmark.cpp
… benchmarks/Vectorization/polybench/MLIRPolybenchGemver.mlir
… benchmarks/Vectorization/polybench/MLIRPolybenchGemverBenchmark.cpp
… benchmarks/Vectorization/polybench/MLIRPolybenchGesummv.mlir
… benchmarks/Vectorization/polybench/MLIRPolybenchGesummvBenchmark.cpp