This PR adds Polybench to vectorization benchmarks.

Most of the Polybench cases uses parametric loop bounds and sizes, which makes it easy to run the benchmark with different dataset sizes. But there are some that uses compile-time constant sizes (including fix-sized local arrays). The MLIR source code here are generated by [Polygeist](https://github.com/llvm/Polygeist) and modified manually to run on different dataset sizes dynamically.

The binary `vectorization-polybench-benchmark` provides two options, `run-benchmark` and `run-validation`.

1. `run-benchmark` runs the benchmark with the specified sizes and prints the execution time of each benchmark. To disable certain dataset sizes, use the `--disable-sizes` option. For example, to disable the mini, large, and extralarge dataset sizes, use the following command:

    ```bash
    ./bin/vectorization-polybench-benchmark run-benchmark --disable-sizes=mini,large,extralarge
    ```

2. `run-validation` runs the benchmark with the specified dataset size and output the result to stdout. The format of output data is the same as the original Polybench suite. So this output can be used to validate the correctness of this port.

    ```bash
    ./bin/vectorization-polybench-benchmark run-validation mini
    ```

    Note that the output can be very large, so it is recommended to redirect the output to a file and compare it with the original Polybench output later.

A benchmarking output on RISC-V platform would be like:

```
Disabled dataset sizes: [ extralarge large mini ]
2025-01-14T19:22:37+08:00
Running ./bin/vectorization-polybench-benchmark
Run on (8 X 1600 MHz CPU s)
CPU Caches:
  L1 Instruction 32 KiB (x8)
  L1 Data 32 KiB (x8)
  L2 Unified 512 KiB (x2)
Load Average: 2.35, 2.09, 2.04
----------------------------------------------------------------
Benchmark                      Time             CPU   Iterations
----------------------------------------------------------------
2mm-small                   2.41 ms         2.41 ms          291
2mm-medium                   186 ms          186 ms            4
3mm-small                   3.78 ms         3.78 ms          185
3mm-medium                   371 ms          371 ms            2
adi-small                   13.8 ms         13.8 ms           51
adi-medium                   530 ms          530 ms            1
atax-small                 0.080 ms        0.080 ms         8720
atax-medium                0.902 ms        0.906 ms          772
bicg-small                 0.209 ms        0.209 ms         3346
bicg-medium                 2.38 ms         2.38 ms          294
cholesky-small              2.08 ms         2.08 ms          337
cholesky-medium              112 ms          112 ms            6
correlation-small           2.55 ms         2.55 ms          273
correlation-medium           117 ms          117 ms            6
covariance-small            2.51 ms         2.51 ms          276
covariance-medium            117 ms          117 ms            6
deriche-small               1.42 ms         1.42 ms          493
deriche-medium              23.2 ms         23.2 ms           30
doitgen-small               2.76 ms         2.76 ms          254
doitgen-medium              45.2 ms         45.2 ms           16
durbin-small               0.107 ms        0.107 ms         6508
durbin-medium               1.14 ms         1.14 ms          615
fdtd-2d-small               2.57 ms         2.57 ms          272
fdtd-2d-medium              74.7 ms         74.6 ms            9
floyd-warshall-small        30.7 ms         30.7 ms           23
floyd-warshall-medium        686 ms          686 ms            1
gemm-small                  1.01 ms         1.01 ms          691
gemm-medium                 25.8 ms         25.8 ms           27
gemver-small               0.234 ms        0.234 ms         2995
gemver-medium               9.26 ms         9.26 ms           75
gesummv-small              0.121 ms        0.121 ms         5793
gesummv-medium             0.920 ms        0.923 ms          758
gramschmidt-small           2.99 ms         2.99 ms          234
gramschmidt-medium           154 ms          154 ms            5
heat-3d-small               6.56 ms         6.56 ms          107
heat-3d-medium               241 ms          241 ms            3
jacobi-1d-small            0.028 ms        0.028 ms        25294
jacobi-1d-medium           0.203 ms        0.203 ms         3445
jacobi-2d-small             3.70 ms         3.69 ms          189
jacobi-2d-medium             198 ms          198 ms            4
lu-small                    4.02 ms         4.01 ms          174
lu-medium                    673 ms          673 ms            1
ludcmp-small                3.86 ms         3.86 ms          181
ludcmp-medium                642 ms          641 ms            1
mvt-small                  0.147 ms        0.147 ms         4766
mvt-medium                  8.15 ms         8.15 ms           86
nussinov-small              4.92 ms         4.92 ms          142
nussinov-medium              253 ms          253 ms            3
seidel-2d-small             17.2 ms         17.2 ms           41
seidel-2d-medium             520 ms          520 ms            1
symm-small                  2.44 ms         2.44 ms          287
symm-medium                  113 ms          113 ms            6
syr2k-small                 3.48 ms         3.48 ms          201
syr2k-medium                 119 ms          119 ms            6
syrk-small                  2.08 ms         2.08 ms          336
syrk-medium                 63.3 ms         63.3 ms           11
trisolv-small              0.054 ms        0.054 ms        12922
trisolv-medium             0.830 ms        0.834 ms          838
trmm-small                 0.948 ms        0.949 ms          737
trmm-medium                 51.9 ms         51.9 ms           13
```