@JuniMay 2025-01-18
I refactored the benchmark registration methods to reuse the google benchmark framework as much as possible and added `-O3` kernels to the benchmarks. Now the benchmarks can be run with the following command:

```bash
./bin/vectorization-polybench-benchmark --benchmark_filter=".*/.*/[12]" --verification-dataset-size=mini
```

which runs all Polybench cases on `small` and `medium` dataset sizes (corresponding to regex `[12]`), and then verify the results of optimized version on `mini` dataset size.

The filter can also be used to run only certain Polybench cases.

Additionally, a simple python script is added to help verify the correctness of this port. Usages of it are placed under the _Verifying the Benchmark Results_ section in README.

A sample of new benchmarking output on RISC-V platform:

```bash
$ ./bin/vectorization-polybench-benchmark --benchmark_filter="MLIRPolybench2mm/.*/[012]" --verification-dataset-size=mini
2025-01-18T21:20:06+08:00
Running ./bin/vectorization-polybench-benchmark
Run on (8 X 1600 MHz CPU s)
CPU Caches:
  L1 Instruction 32 KiB (x8)
  L1 Data 32 KiB (x8)
  L2 Unified 512 KiB (x2)
Load Average: 2.46, 2.66, 2.42
---------------------------------------------------------------------
Benchmark                           Time             CPU   Iterations
---------------------------------------------------------------------
MLIRPolybench2mm/scalar/0       0.535 ms        0.535 ms         1307
MLIRPolybench2mm/scalar/1        12.0 ms         12.0 ms           58
MLIRPolybench2mm/scalar/2         653 ms          651 ms            1
MLIRPolybench2mm/autovec/0      0.108 ms        0.108 ms         6476
MLIRPolybench2mm/autovec/1       2.42 ms         2.41 ms          290
MLIRPolybench2mm/autovec/2        185 ms          185 ms            4
------------------------------------------------
autovec 2mm-mini PASS
autovec 3mm-mini PASS
autovec adi-mini PASS
autovec atax-mini PASS
autovec bicg-mini (s) PASS
autovec bicg-mini (q) PASS
autovec cholesky-mini PASS
autovec correlation-mini PASS
autovec covariance-mini PASS
autovec deriche-mini PASS
autovec doitgen-mini PASS
autovec durbin-mini PASS
autovec fdtd-2d-mini (ex) PASS
autovec fdtd-2d-mini (ey) PASS
autovec fdtd-2d-mini (hz) PASS
autovec floyd-warshall-mini PASS
autovec gemm-mini PASS
autovec gemver-mini PASS
autovec gesummv-mini PASS
autovec gramschmidt-mini (R) PASS
autovec gramschmidt-mini (Q) PASS
autovec heat-3d-mini PASS
autovec jacobi-1d-mini PASS
autovec jacobi-2d-mini PASS
autovec lu-mini PASS
autovec ludcmp-mini PASS
autovec mvt-mini (x1) PASS
autovec mvt-mini (x2) PASS
autovec nussinov-mini PASS
autovec seidel-2d-mini PASS
autovec symm-mini PASS
autovec syr2k-mini PASS
autovec syrk-mini PASS
autovec trisolv-mini PASS
autovec trmm-mini PASS
------------------------------------------------
```