@xlinsist 2025-01-19
> The filter can also be used to run only certain Polybench cases.

Commands like `./bin/vectorization-polybench-benchmark --benchmark_filter="MLIRPolybench2mm/.*/[012]"` can be added to the README as an example usage.