@xlinsist 2025-01-19
Though some parts of the code require modification and refactoring, this PR is overall well-completed. It also includes verification scripts and test results, which I have reviewed. Well done!

Out of curiosity, since polybench includes many cases, have you sought assistances from hand-written automated/semi-automated scripts, to help improve your efficiency?"