@JuniMay 2025-01-19
Thanks, and I'm working on the modification of the required parts.

And yes, I tried to automate the porting process from the beginning of this PR. The script provided under the `polybench` directory is used to generate all MLIR codes with Polygeist, and to verify the results conveniently. I also tried using LLMs to modify MLIR files and generate the benchmarking code, but it turned out to be a failure (I tried to prompt engineer Qwen, 4o, DeepSeek, etc., but failed to generate expected code).

There are some repetitive parts that can be simplified with macros, like `_scalar` and `_autovec` method names. However, considering the different kernel signatures and that we might want to optimize a single benchmark case with specialized pipelines (which requires flexibility over simplicity) I kept them the same as the current implementation.