@xlinsist 2025-01-19
> The script provided under the polybench directory is used to generate all MLIR codes with Polygeist, and to verify the results conveniently. I also tried using LLMs to modify MLIR files and generate the benchmarking code, but it turned out to be a failure (I tried to prompt engineer Qwen, 4o, DeepSeek, etc., but failed to generate expected code).
I see.

> considering the different kernel signatures and that we might want to optimize a single benchmark case with specialized pipelines

It makes sense. We also need to beware of the trap of oversimplifying code.